requirements.txt
numpy>=1.24
Pillow>=10.0
openai>=1.40.0
tqdm>=4.66
open_clip_torch>=2.24.0
